added new menu options
