+++ requirements.txt
@@ -8,3 +8,7 @@ stellar_base==1.1.2.0
 gql==3.4.0
 cbor2==5.4.3
 bech32==1.2.0
+crypto==1.4.1
+qrcode==7.3.1
+Pillow==9.2.0
+inquirer==2.9.0
